// core/src/generated/api/PerspectiveHandle.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DecoratedNeighbourhoodExpression } from "./DecoratedNeighbourhoodExpression";
import type { PerspectiveState } from "./PerspectiveState";

export type PerspectiveHandle = { uuid: string, name: string | null, neighbourhood: DecoratedNeighbourhoodExpression | null, sharedUrl: string | null, state: PerspectiveState, owners: Array<string> | null, };