// core/src/generated/api/DecoratedNeighbourhoodExpression.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DecoratedExpressionProof } from "./DecoratedExpressionProof";
import type { Neighbourhood } from "./Neighbourhood";

export type DecoratedNeighbourhoodExpression = { author: string, data: Neighbourhood, proof: DecoratedExpressionProof, timestamp: string, };